move view models to sepaate module and clean code
